As a follow-up to #276. The `int32` version is useful because R uses 32-bit integers to represent boolean (i.e., logical) arrays. This results in a significant speedup in boolean conversion!

@WillAyd: I updated a few things that you *just* added (Sorry! 😬 ):

- I changed `Bitmap` -> `Bits` and removed `Unsafe` to make it more consistent with the other functions that accept `const uint8_t* bits`
- I updated the test function so that it tests both the int32 and int8 types at once

Before this PR:

``` r
library(nanoarrow)

lgls <- nanoarrow:::vec_gen(logical(), 1e6)
bool_array <- as_nanoarrow_array(lgls)
bool_array_arrow <- arrow::as_arrow_array(bool_array)

bench::mark(
  convert_array(bool_array, logical()),
  as.vector(bool_array_arrow),
  as.logical(lgls)
)
#> # A tibble: 3 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 convert_array(bool_array, logical()) 556µs  749µs    1.33e3    3.82MB     156.
#> 2 as.vector(bool_array_arrow)          558µs  780µs    1.30e3    3.82MB     144.
#> 3 as.logical(lgls)                         0    1ns    2.28e8        0B       0

bench::mark(
  convert_array(bool_array, integer()),
  as.integer(lgls)
)
#> # A tibble: 2 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 convert_array(bool_array, integer()) 733µs  912µs     1093.    3.81MB     167.
#> 2 as.integer(lgls)                     615µs  788µs     1273.    3.81MB     182.
```

After this PR:

``` r
library(nanoarrow)

lgls <- nanoarrow:::vec_gen(logical(), 1e6)
bool_array <- as_nanoarrow_array(lgls)
bool_array_arrow <- arrow::as_arrow_array(bool_array)

bench::mark(
  convert_array(bool_array, logical()),
  as.vector(bool_array_arrow),
  as.logical(lgls)
)
#> # A tibble: 3 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 convert_array(bool_array, logical()) 105µs  308µs    3.21e3    3.83MB     367.
#> 2 as.vector(bool_array_arrow)          559µs  772µs    1.30e3    3.82MB     143.
#> 3 as.logical(lgls)                         0      0    5.87e8        0B       0

bench::mark(
  convert_array(bool_array, integer()),
  as.integer(lgls)
)
#> # A tibble: 2 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 convert_array(bool_array, integer()) 104µs  310µs     3181.    3.81MB     423.
#> 2 as.integer(lgls)                     615µs  784µs     1278.    3.81MB     142.
```

<sup>Created on 2023-08-17 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>